requirements.txt
python-chess
pandas
cairosvg
fastai
jupyter